Paso a paso de double

+++ 3er intento/decimales/src/App.java
@@ -3,13 +3,13 @@ public class App {
         double salario = 1250.56;
         System.out.println(salario);
 
-        double edad = 28;
-        double salarioMitad = salario /2;
+        int edad = 28;
+        double salarioMitad = salario / 2;
         System.out.println(salarioMitad);
 
         //Primero se declara la variable 
-        int division;
-        division = 1250/ 3;
+        double division;
+        division = 1250.0 / 3.0;
         System.out.println(division);
     }
 }
